testing multiple y-axis points for hbar_stack
Per the developer, this functionality does not exist.

Turns out that the current functionality for the stacked horizontal bar chart I was using only allows for bar compilations starting at the 0 value.

I made a GitHub issue to request the feature.
In the meantime, the developer suggested that I use another method to create the visualization. Essentially, I'd be making multiple bar graphs and concatenating them manually. I'm not sure how much longer this'll take under the hood, but it seems to be the only option as of now.